feat: Refactor move computing of tabs to Utils file + Set current active tab to first of available
